tests/infrastructure/requirements.txt
# Test dependencies for SVOA Lea Infrastructure Tests
# Python 3.9+ required

# Core testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-timeout==2.2.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# AWS SDK and tools
boto3==1.34.14
botocore==1.34.14
moto==4.2.12  # AWS service mocking
aws-sam-cli==1.105.0

# Kubernetes
kubernetes==29.0.0
kubectl==0.16.0
kube-hunter==0.6.8  # Security testing

# Infrastructure as Code
terraform==1.0.3
python-hcl2==4.3.0  # Parse HCL files
checkov==3.1.38  # IaC security scanning
tfsec==1.28.4
terrascan==1.18.11

# Container and Docker
docker==7.0.0
container-structure-test==1.11.0
dive==0.11.0  # Docker image analysis

# Security scanning
bandit==1.7.5
safety==3.0.1
trivy==0.48.1
snyk==1.1266.0
semgrep==1.52.0
gitleaks==8.18.1

# Cryptography and secrets
cryptography==41.0.7
hvac==2.0.0  # HashiCorp Vault
python-jose==3.3.0
PyJWT==2.8.0

# Network and SSL
requests==2.31.0
pyOpenSSL==23.3.0
ssl-checker==2.0.7
dnspython==2.4.2

# YAML and JSON processing
PyYAML==6.0.1
jsonschema==4.20.0
jmespath==1.0.1
yq==3.2.3

# GitOps and CI/CD
python-gitlab==4.2.0
PyGithub==2.1.1
python-jenkins==1.8.2

# Monitoring and metrics
prometheus-client==0.19.0
datadog==0.47.0
grafana-client==3.6.0

# Performance testing
locust==2.19.1
pytest-benchmark==4.0.0

# EU/GDPR compliance tools
gdpr-check==0.1.0
privacy-api==0.2.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1
pylint==3.0.3
isort==5.13.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
myst-parser==2.0.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0  # Better console output
tabulate==0.9.0
tqdm==4.66.1
ipaddress==1.0.23

# Swedish-specific dependencies
svenska-validator==0.1.0  # Swedish data validation
personnummer==3.1.0  # Swedish personal number validation